Fixed several bugs in ptmsearch. In addition, revised ps_align.cpp speedup

+++ src/test/test_diag_filter.cpp
@@ -45,18 +45,16 @@ int zero_ptm_process(int argc, char* argv[]) {
       generateShuffleDb(ori_db_file_name, db_file_name);
     }
 
-    int start_s = clock();
+    long start_s = clock();
 
-    /*
     std::cout << "Fast filtering starts " << std::endl;
     PtmFastFilterMngPtr filter_mng_ptr 
         = PtmFastFilterMngPtr(new PtmFastFilterMng(prsm_para_ptr, "FILTER"));
     PtmFastFilterProcessorPtr filter_processor = PtmFastFilterProcessorPtr(new PtmFastFilterProcessor(filter_mng_ptr));
     filter_processor->process();
     filter_processor = nullptr;
-    */
 
-    int stop_s = clock();
+    long stop_s = clock();
     std::cout << std::endl << "Running time: " << (stop_s-start_s) / double(CLOCKS_PER_SEC)  << " seconds " << std::endl;
 
     std::cout << "Outputting table starts " << std::endl;
